Pricesdb

+++ bot/handlers/other.py
@@ -1,9 +1,9 @@
 from aiogram import Dispatcher, Bot
 from aiogram.types import Message
 
-from bot.db import read_user, add_link, delete_link, update_last_command
+from bot.db import read_user, add_link, delete_link, update_last_command, update_price
 from bot.db.main import User, Link
-from bot.misc import logger, BAD_MSG
+from bot.misc import logger, BAD_MSG, BAD_URL, GOOD_URL, GOOD_DELETE
 
 
 async def other_messages(msg: Message) -> None:
@@ -18,28 +18,34 @@ async def other_messages(msg: Message) -> None:
         case _:
             bot: Bot = msg.bot
             await bot.send_message(msg.from_user.id, BAD_MSG)
+            await update_last_command(data=User(id=msg.from_user.id, command=""))
 
 
 async def __add_link(msg: Message) -> None:
-    try:
-        await add_link(data=Link(id=msg.from_user.id, url=msg.text, price=0))
-    except Exception as ex:
-        logger.error(ex)
+    if msg.text.startswith("https://www.ozon.ru/product/"):
+        try:
+            await add_link(link=Link(id=msg.from_user.id, url=msg.text, price=0))
+            await update_price(link=Link(id=msg.from_user.id, url=msg.text, price=0))
+        except Exception as ex:
+            logger.error(ex)
+        else:
+            bot: Bot = msg.bot
+            await bot.delete_message(chat_id=msg.from_user.id, message_id=msg.message_id)
+            await bot.send_message(chat_id=msg.from_user.id, text=GOOD_URL)
     else:
         bot: Bot = msg.bot
-        await bot.delete_message(chat_id=msg.from_user.id, message_id=msg.message_id)
-        await bot.send_message(chat_id=msg.from_user.id, text="Ваш url успешно добавлен в подписку!")
+        await bot.send_message(chat_id=msg.from_user.id, text=BAD_URL)
 
 
 async def __delete_link(msg: Message) -> None:
     try:
-        await delete_link(data=Link(id=msg.from_user.id, url=msg.text, price=0))
+        await delete_link(link=Link(id=msg.from_user.id, url=msg.text, price=0))
     except Exception as ex:
         logger.error(ex)
     else:
         bot: Bot = msg.bot
         await bot.delete_message(chat_id=msg.from_user.id, message_id=msg.message_id)
-        await bot.send_message(chat_id=msg.from_user.id, text="Ваш url успешно удален из подписки!")
+        await bot.send_message(chat_id=msg.from_user.id, text=GOOD_DELETE)
 
 
 def register_other_handlers(dp: Dispatcher) -> None:
